Move of input package to the odalic root package.
