+++ odalic/src/main/java/cz/cuni/mff/xrg/odalic/api/rest/resources/FeedbackResource.java
@@ -17,7 +17,7 @@ import org.springframework.stereotype.Component;
 import com.google.common.base.Preconditions;
 
 import cz.cuni.mff.xrg.odalic.feedbacks.Feedback;
-import cz.cuni.mff.xrg.odalic.feedbacks.input.Input;
+import cz.cuni.mff.xrg.odalic.input.Input;
 import cz.cuni.mff.xrg.odalic.tasks.feedbacks.FeedbackService;
 
 @Component
