+++ odalic/src/main/java/cz/cuni/mff/xrg/odalic/input/CsvInputParser.java
@@ -1,4 +1,4 @@
-package cz.cuni.mff.xrg.odalic.feedbacks.input;
+package cz.cuni.mff.xrg.odalic.input;
 
 import java.io.IOException;
 import java.io.InputStream;
