+++ odalic/src/main/java/cz/cuni/mff/xrg/odalic/input/InputToTableAdapter.java
@@ -1,4 +1,4 @@
-package cz.cuni.mff.xrg.odalic.feedbacks.input;
+package cz.cuni.mff.xrg.odalic.input;
 
 import uk.ac.shef.dcs.sti.core.model.Table;
