+++ odalic/src/main/java/cz/cuni/mff/xrg/odalic/input/ListsBackedInput.java
@@ -1,4 +1,4 @@
-package cz.cuni.mff.xrg.odalic.feedbacks.input;
+package cz.cuni.mff.xrg.odalic.input;
 
 import java.io.Serializable;
 import java.util.List;
