+++ odalic/src/main/java/cz/cuni/mff/xrg/odalic/tasks/feedbacks/MemoryOnlyFeedbackService.java
@@ -7,11 +7,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import com.google.common.base.Preconditions;
 
 import cz.cuni.mff.xrg.odalic.feedbacks.Feedback;
-import cz.cuni.mff.xrg.odalic.feedbacks.input.Input;
-import cz.cuni.mff.xrg.odalic.feedbacks.input.CsvConfiguration;
-import cz.cuni.mff.xrg.odalic.feedbacks.input.CsvInputParser;
 import cz.cuni.mff.xrg.odalic.files.File;
 import cz.cuni.mff.xrg.odalic.files.FileService;
+import cz.cuni.mff.xrg.odalic.input.CsvConfiguration;
+import cz.cuni.mff.xrg.odalic.input.CsvInputParser;
+import cz.cuni.mff.xrg.odalic.input.Input;
 import cz.cuni.mff.xrg.odalic.tasks.configurations.Configuration;
 import cz.cuni.mff.xrg.odalic.tasks.configurations.ConfigurationService;
