+++ odalic/src/main/resources/spring/applicationContext.xml
@@ -13,13 +13,13 @@
 
 	<context:component-scan base-package="cz.cuni.mff.xrg.odalic.*" />
 
-	<bean id="inputBuilder" class="cz.cuni.mff.xrg.odalic.feedbacks.input.ListsBackedInputBuilder"
+	<bean id="inputBuilder" class="cz.cuni.mff.xrg.odalic.input.ListsBackedInputBuilder"
 		scope="prototype" />
-	<bean id="inputToTableAdapter" class="cz.cuni.mff.xrg.odalic.feedbacks.input.DefaultInputToTableAdapter"
+	<bean id="inputToTableAdapter" class="cz.cuni.mff.xrg.odalic.input.DefaultInputToTableAdapter"
 		scope="prototype" />
 	
 	<bean id="csvInputParser"
-		class="cz.cuni.mff.xrg.odalic.feedbacks.input.DefaultCsvInputParser"
+		class="cz.cuni.mff.xrg.odalic.input.DefaultCsvInputParser"
 		scope="prototype" />
 	<bean id="annotationToResultAdapter"
 		class="cz.cuni.mff.xrg.odalic.tasks.results.DefaultAnnotationToResultAdapter"
